Update: Specification of WS request Objects & Support tweaks for Windows

+++ src/Core/Parallel/Parallel.php
@@ -44,14 +44,15 @@ use ReflectionClass;
 use Revolt\EventLoop;
 use Throwable;
 
-use function count;
-use function dirname;
-use function file_exists;
-use function intval;
-use function is_int;
 use function Co\cancel;
 use function Co\defer;
 use function Co\onSignal;
+use function count;
+use function dirname;
+use function file_exists;
+use function getmypid;
+use function intval;
+use function is_int;
 use function posix_getpid;
 use function posix_kill;
 use function preg_match;
@@ -59,6 +60,7 @@ use function shell_exec;
 use function strval;
 
 use const SIGUSR2;
+use const PHP_OS_FAMILY;
 
 /**
  * 2024-08-07
@@ -174,11 +176,24 @@ class Parallel extends LibraryAbstract
         $this->counterRuntime = new Runtime();
         $this->counterFuture = $this->counterRuntime->run(static function ($channel, $eventScalar) {
             $eventScalar(fn () => $eventScalar->wait());
-            $processId = posix_getpid();
+            /**
+             * @compatible:Windows
+             */
+            if (PHP_OS_FAMILY === 'Windows') {
+                $processId = getmypid();
+            } else {
+                $processId = posix_getpid();
+            }
             $count = 0;
             while($number = $channel->recv()) {
                 $eventScalar->set($count += $number);
                 if($number > 0) {
+                    /**
+                     * @compatible:Windows
+                     */
+                    if (PHP_OS_FAMILY === 'Windows') {
+                        break;
+                    }
                     posix_kill($processId, SIGUSR2);
                 } elseif($count === -1) {
                     break;
